+++ mspm0-data-types/src/lib.rs
@@ -43,10 +43,10 @@ pub struct Chip {
     /// Number of options for VRSEL of the ADC peripheral.
     ///
     /// This is requried because we use a single adc_v1 pac for all chips.
-    pub adc_vrsel: u32,
+    pub adc_vrsel: u8,
 
     /// Number adc analog channels available on the chip.
-    pub adc_analog_chan: u32,
+    pub adc_analog_chan: u8,
 
     /// ADC channels per ADC peripheral available on the chip.
     pub adc_channels: BTreeMap<u32, BTreeMap<u32, AdcChannel>>,
